+++ src/main/java/Controller/ControllerShop.java
@@ -79,13 +79,13 @@ public class ControllerShop {
     }
 
     private boolean switchController() {
-        Integer inputInt = input.getIntInput("Please provide option.");
+        Integer inputInt =  1;//input.getIntInput("Please provide option.");
         switch (inputInt) {
             case 0:
                 return false;
             case 1:
                 chooseProduct(searchProducts());
-                break;
+                //break;
             case 2:
                 view.print(controllerClient.getBasket());
                 break;
@@ -93,7 +93,8 @@ public class ControllerShop {
                 executeOrder();
                 break;
         }
-        return true;
+        return false;
+        //return true;
     }
 
     public void run() {
