+++ src/main/java/View/AbstractView.java
@@ -7,7 +7,6 @@ import java.util.List;
 import java.util.TreeMap;
 
 import java.util.HashMap;
-import java.util.Map;
 
 
 public abstract class AbstractView {
@@ -18,18 +17,9 @@ public abstract class AbstractView {
 
     public abstract void print(TreeMap<Product, Integer> productsList);
 
-    public void print(String[] menuContent, String label) {
-        System.out.println(label);
-        for (String menuItem : menuContent) {
-            System.out.println(menuItem);
-        }
-    }
+    public abstract void print(String[] menuContent, String label);
 
-    public void print(HashMap<Product, Integer> basket) {
-        for (Map.Entry<Product, Integer> product : basket.entrySet()) {
-            System.out.println(product.getKey().toString() + " " + product.getValue());
-        }
-    }
+    public abstract void print(HashMap<Product, Integer> basket);
 }
