+++ src/main/java/View/TerminalView.java
@@ -2,6 +2,8 @@ package View;
 
 import Model.Product;
 import Model.UserAbstract;
+
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
@@ -26,6 +28,21 @@ public class TerminalView extends AbstractView {
     public void print(TreeMap<Product, Integer> productsList) {
         for (Map.Entry<Product,Integer> product: productsList.entrySet()
         ) {
+            System.out.println(product.getKey().toString() + " " + product.getValue());
+        }
+    }
+
+    @Override
+    public void print(String[] menuContent, String label) {
+        System.out.println(label);
+        for (String menuItem : menuContent) {
+            System.out.println(menuItem);
+        }
+    }
+
+    @Override
+    public void print(HashMap<Product, Integer> basket) {
+        for (Map.Entry<Product, Integer> product : basket.entrySet()) {
             System.out.println(product.getKey().toString() + " " + product.getValue());
         }
     }
